Feature BL-15043 linting and static analysis

+++ src/CpmsCommon/Utility/Util.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace CpmsCommon\Utility;
 
 /**
@@ -11,7 +12,7 @@ class Util
     /**
      * Method to append any additional data to the clientUrl
      *
-     * @param        $url
+     * @param string $url
      * @param array  $requiredParams
      * @param string $separator
      *
@@ -35,14 +36,14 @@ class Util
     }
 
     /**
-     * @param $path
+     * @param string $path
      *
      * @return bool
      */
     public static function deleteDir($path)
     {
-        if (is_dir($path) === true) {
-            $files = array_diff(scandir($path), array('.', '..'));
+        if (is_dir($path) === true && $fileList = scandir($path)) {
+            $files = array_diff($fileList, array('.', '..'));
 
             foreach ($files as $file) {
                 self::deleteDir(realpath($path) . '/' . $file);
